terminfo: improve xterm definition
Add xterm-256color as an alias for xterm.

Change xterm key sequences (up, down, left, right, etc.) to the standard mode
sequences instead of application mode. Typically in terminfo definitions, the
assumption is the application will use the smkx/rmkx sequences to put the
terminal in application mode. However, ScarletDME doesn't seem to do that, so
we need to accept the standard mode key sequences for them to be detected
properly.