Add $SVLISTS/$SAVEDLISTS back to VOC
This un-does commit 9f6ca2b. The message in that commit says that
$SAVEDLISTS will be created automatically on first use, however
that does not appear to be correct. When creating a new account,
it is created at account creation time. However, if you use a
command such as UPDATE.RECORDS in the QMSYS account, the command
fails due to being unable to open $SAVEDLISTS. Since $SAVEDLISTS
is created in every new account from the start by the
CREATE.ACCOUNT command, it should also already exist in the QMSYS
account.